+++ src/main/java/com/langqiao/sharding/jdbc/algorithm/UserSingleKeyTableShardingAlgorithm.java
@@ -7,12 +7,12 @@ import com.dangdang.ddframe.rdb.sharding.api.ShardingValue;
 import com.dangdang.ddframe.rdb.sharding.api.strategy.table.SingleKeyTableShardingAlgorithm;
 import com.google.common.collect.Range;
 
-public class UserSingleKeyTableShardingAlgorithm implements SingleKeyTableShardingAlgorithm<Integer>{
+public class UserSingleKeyTableShardingAlgorithm implements SingleKeyTableShardingAlgorithm<Long>{
 
     /**
      * sql 中 = 操作时，table的映射
      */
-    public String doEqualSharding(Collection<String> tableNames, ShardingValue<Integer> shardingValue) {
+    public String doEqualSharding(Collection<String> tableNames, ShardingValue<Long> shardingValue) {
         for (String each : tableNames) {
             if (each.endsWith(shardingValue.getValue() % 3 + "")) {
                 return each;
@@ -24,9 +24,9 @@ public class UserSingleKeyTableShardingAlgorithm implements SingleKeyTableShardi
     /**
      * sql 中 in 操作时，table的映射
      */
-    public Collection<String> doInSharding(Collection<String> tableNames, ShardingValue<Integer> shardingValue) {
+    public Collection<String> doInSharding(Collection<String> tableNames, ShardingValue<Long> shardingValue) {
         Collection<String> result = new LinkedHashSet<String>(tableNames.size());
-        for (Integer value : shardingValue.getValues()) {
+        for (Long value : shardingValue.getValues()) {
             for (String tableName : tableNames) {
                 if (tableName.endsWith(value % 3 + "")) {
                     result.add(tableName);
@@ -40,10 +40,10 @@ public class UserSingleKeyTableShardingAlgorithm implements SingleKeyTableShardi
      * sql 中 between 操作时，table的映射
      */
     public Collection<String> doBetweenSharding(Collection<String> tableNames,
-            ShardingValue<Integer> shardingValue) {
+            ShardingValue<Long> shardingValue) {
         Collection<String> result = new LinkedHashSet<String>(tableNames.size());
-        Range<Integer> range = (Range<Integer>) shardingValue.getValueRange();
-        for (Integer i = range.lowerEndpoint(); i <= range.upperEndpoint(); i++) {
+        Range<Long> range = (Range<Long>) shardingValue.getValueRange();
+        for (Long i = range.lowerEndpoint(); i <= range.upperEndpoint(); i++) {
             for (String each : tableNames) {
                 if (each.endsWith(i % 3 + "")) {
                     result.add(each);
